use int instead of long
git-svn-id: file:///source/coreland/corelib/trunk@2192 6e7d7e42-f3d6-da11-80d8-00d0b73f4485

+++ fmt_float.c
@@ -38,7 +38,7 @@ static inline float powf(float x, float y) { return (float) pow(x, y); }
 #endif
 
 union real {
-  unsigned long n;
+  unsigned int n;
   float f;
 } real;
 typedef unsigned int fmt_func(char *, unsigned long);
